moving reusable components into the shared library
