+++ clients/interview_component/src/interview/components/InterviewCard.tsx
@@ -8,7 +8,7 @@ import { TooltipProvider } from '@/components/ui/tooltip'
 import { MinimalTiptapEditor } from '@/components/minimal-tiptap/minimal-tiptap'
 import { useCoursePhaseStore } from '../zustand/useCoursePhaseStore'
 import { useParticipationStore } from '../zustand/useParticipationStore'
-import { useUpdateCoursePhaseParticipation } from '../hooks/useUpdateCoursePhaseParticipation'
+import { useUpdateCoursePhaseParticipation } from '@/hooks/useUpdateCoursePhaseParticipation'
 import type { InterviewQuestion } from '../interfaces/InterviewQuestion'
 import type { InterviewAnswer } from '../interfaces/InterviewAnswer'
 import { Input } from '@/components/ui/input'
