+++ clients/interview_component/src/interview/components/InterviewTimesDialog.tsx
@@ -11,7 +11,7 @@ import { Button } from '@/components/ui/button'
 import { Separator } from '@/components/ui/separator'
 import { Clock, Plus, Trash } from 'lucide-react'
 import { useCoursePhaseStore } from '../zustand/useCoursePhaseStore'
-import { useUpdateCoursePhaseMetaData } from '../hooks/useUpdateCoursePhaseMetaData'
+import { useUpdateCoursePhaseMetaData } from '@/hooks/useUpdateCoursePhaseMetaData'
 import { useParticipationStore } from '../zustand/useParticipationStore'
 import { Input } from '@/components/ui/input'
 import {
