+++ clients/shared_library/hooks/useUpdateCoursePhaseParticipation.tsx
@@ -1,13 +1,13 @@
 import { useToast } from '@/hooks/use-toast'
 import { useMutation, UseMutationResult, useQueryClient } from '@tanstack/react-query'
 import { useParams } from 'react-router-dom'
-import { updateCoursePhase } from '../network/mutations/updateCoursePhase'
-import { UpdateCoursePhase } from '@tumaet/prompt-shared-state'
+import { UpdateCoursePhaseParticipation } from '@tumaet/prompt-shared-state'
+import { updateCoursePhaseParticipation } from '@/network/mutations/updateCoursePhaseParticipationMetaData'
 
-export const useUpdateCoursePhaseMetaData = (): UseMutationResult<
+export const useUpdateCoursePhaseParticipation = (): UseMutationResult<
   string | undefined,
   Error,
-  UpdateCoursePhase,
+  UpdateCoursePhaseParticipation,
   unknown
 > => {
   const { phaseId } = useParams<{ phaseId: string }>()
@@ -16,20 +16,20 @@ export const useUpdateCoursePhaseMetaData = (): UseMutationResult<
   const { toast } = useToast()
 
   const mutation = useMutation({
-    mutationFn: (coursePhase: UpdateCoursePhase) => {
-      return updateCoursePhase(coursePhase)
+    mutationFn: (coursePhaseParticipation: UpdateCoursePhaseParticipation) => {
+      return updateCoursePhaseParticipation(coursePhaseParticipation)
     },
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['course_phase', phaseId] })
+      queryClient.invalidateQueries({ queryKey: ['participants', phaseId] })
       toast({
         title: 'Success',
-        description: 'Successfully updated the interview questions.',
+        description: 'Successfully updated the course participation.',
       })
     },
     onError: () => {
       toast({
         title: 'Error',
-        description: 'Failed to update the interview questions.',
+        description: 'Failed to update the course participation.',
         variant: 'destructive',
       })
     },
